Separate date component without SSR
